Moved formatTimestamp function to global utils file
