+++ src/utils.tsx
@@ -1,0 +1,32 @@
+/**
+ * Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * Converts a timestamp into a consistent and formatted firebase-log date
+ */
+export const formatTimestamp = (timestamp: number) => {
+  const date = new Date(timestamp);
+  const segments = [date.getHours(), date.getMinutes(), date.getSeconds()];
+
+  return segments
+    .map((segment) => {
+      const number = segment.toString();
+
+      if (number.length === 2) return number;
+      else return `0${number}`;
+    })
+    .join(':');
+};
